完善保存

+++ src/main/java/save/edit/model/SaveModel.java
@@ -25,9 +25,9 @@ public class SaveModel {
 	private int m_fTalkRange;
 	private float m_fStopRange;
 	private float m_fWalkRange;
-	private float m_fNowX;
-	private float m_fNowY;
-	private float m_fNowZ;
+	private double m_fNowX;
+	private double m_fNowY;
+	private double m_fNowZ;
 	private int m_iNowTime;
 	private int m_fDir;
 	private float m_fNowTime;
@@ -167,22 +167,22 @@ public class SaveModel {
 	public void setM_fWalkRange(float m_fWalkRange) {
 		this.m_fWalkRange = m_fWalkRange;
 	}
-	public float getM_fNowX() {
+	public double getM_fNowX() {
 		return m_fNowX;
 	}
-	public void setM_fNowX(float m_fNowX) {
+	public void setM_fNowX(double m_fNowX) {
 		this.m_fNowX = m_fNowX;
 	}
-	public float getM_fNowY() {
+	public double getM_fNowY() {
 		return m_fNowY;
 	}
-	public void setM_fNowY(float m_fNowY) {
+	public void setM_fNowY(double m_fNowY) {
 		this.m_fNowY = m_fNowY;
 	}
-	public float getM_fNowZ() {
+	public double getM_fNowZ() {
 		return m_fNowZ;
 	}
-	public void setM_fNowZ(float m_fNowZ) {
+	public void setM_fNowZ(double m_fNowZ) {
 		this.m_fNowZ = m_fNowZ;
 	}
 	public int getM_iNowTime() {
